cohort_2/requirements.txt
accelerate==1.3.0
anthropic>=0.40.0
bertopic>=0.16.4
braintrust>=0.0.177
bs4>=0.0.2
cohere>=5.11.3
datasets>=3.1.0
faker>=33.1.0
fuzzywuzzy>=0.18.0
instructor[google-generativeai]>=1.7.2
ipykernel>=6.29.5
ipywidgets>=8.1.5
jsonref>=1.1.0
lancedb>=0.16.0
markdownify>=0.13.1
matplotlib>=3.9.2
modal>=0.65.38
nbformat>=5.10.4
open-clip-torch>=2.29.0
openai>=1.57.0
pandas>=2.2.3
pillow>=11.0.0
pydantic>=2.10.1
python-levenshtein>=0.26.1
pyyaml>=6.0.2
requests>=2.32.3
scikit-learn>=1.6.0
seaborn>=0.13.2
sentence-transformers==3.1.1
streamlit>=1.40.1
streamlit-shortcuts>=0.1.9
tantivy>=0.22.0
transformers==4.45.2 
eval_type_backport==0.2.2